Link io to nomt
