+++ nomt/Cargo.toml
@@ -21,12 +21,18 @@ blake3 = "1.5.1"
 fxhash = "0.2.1"
 dashmap = "5.5.3"
 crossbeam = "0.8.4"
+crossbeam-channel = "0.5.13"
+slab = "0.4.9"
+rand = "0.8.5"
+ahash = "0.8.11"
+im = "15.1.0"
 lru = "0.12.3"
+libc = "0.2.155"
+io-uring = "0.6.4"
 
 [target.'cfg(loom)'.dependencies]
 loom = { version = "0.7", features = ["checkpoint"] }
 
 [dev-dependencies]
-rand = "0.8.5"
 rand_pcg = "0.3.1"
 hex-literal = "0.4"
